backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.12
stripe==10.4.0
